+++ clients/derivatives_trading_coin_futures/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 5.0.0 - 2026-03-09
+
+### Changed (1)
+
+#### REST API
+
+- Modified response for `exchange_information()` (`GET /dapi/v1/exchangeInfo`):
+  - `symbols`.items: property `orderTypes` added
+  - `symbols`.items: property `OrderType` deleted
+  - `symbols`.items: item property `orderTypes` added
+  - `symbols`.items: item property `OrderType` deleted
+
 ## 4.1.1 - 2026-02-25
 
 ### Changed (1)